Move layouts components into components dir
