--- src/layouts/index.ts
@@ -1,2 +1,0 @@
-export * from './error';
-export * from './home';
